backend/requirements.txt
Django>=5.0
djangorestframework
mysqlclient
python-dotenv
django-cors-headers
django-otp